+++ mimesis/enums.py
@@ -15,6 +15,50 @@ Below you can see an example of usage enums in methods of data providers.
 from enum import Enum
 
 
+class Locale(Enum):
+    """This class provides access to the supported locales from one place."""
+
+    CS = "cs"
+    DA = "da"
+    DE = "de"
+    DE_AT = "de-at"
+    DE_CH = "de-ch"
+    EL = "el"
+    EN = "en"
+    EN_AU = "en-au"
+    EN_CA = "en-ca"
+    EN_GB = "en-gb"
+    ES = "es"
+    ES_MX = "es-mx"
+    ET = "et"
+    FA = "fa"
+    FI = "fi"
+    FR = "fr"
+    HU = "hu"
+    IS = "is"
+    IT = "it"
+    JA = "ja"
+    KK = "kk"
+    KO = "ko"
+    NL = "nl"
+    NL_BE = "nl-be"
+    NO = "no"
+    PL = "pl"
+    PT = "pt"
+    PT_BR = "pt-br"
+    RU = "ru"
+    SK = "sk"
+    SV = "sv"
+    TR = "tr"
+    UK = "uk"
+    ZH = "zh"
+    DEFAULT = EN
+
+    @classmethod
+    def values(cls):
+        return [i.value for i in cls.__members__.values()]
+
+
 class PortRange(Enum):
     """Represents port ranges.
