+++ mimesis/locales.py
@@ -1,50 +1,24 @@
 # -*- coding: utf-8 -*-
 
 """This module provides constants for locale-dependent providers."""
-from enum import Enum
+
+from mimesis.enums import Locale
+from mimesis.exceptions import LocaleError
 
 __all__ = ["Locale"]
 
 
-class Locale(Enum):
-    """This class provides access to the supported locales from one place."""
+def validate_locale(locale: Locale) -> Locale:
+    if not locale:
+        locale = Locale.DEFAULT
 
-    CS = "cs"
-    DA = "da"
-    DE = "de"
-    DE_AT = "de-at"
-    DE_CH = "de-ch"
-    EL = "el"
-    EN = "en"
-    EN_AU = "en-au"
-    EN_CA = "en-ca"
-    EN_GB = "en-gb"
-    ES = "es"
-    ES_MX = "es-mx"
-    ET = "et"
-    FA = "fa"
-    FI = "fi"
-    FR = "fr"
-    HU = "hu"
-    IS = "is"
-    IT = "it"
-    JA = "ja"
-    KK = "kk"
-    KO = "ko"
-    NL = "nl"
-    NL_BE = "nl-be"
-    NO = "no"
-    PL = "pl"
-    PT = "pt"
-    PT_BR = "pt-br"
-    RU = "ru"
-    SK = "sk"
-    SV = "sv"
-    TR = "tr"
-    UK = "uk"
-    ZH = "zh"
-    DEFAULT = EN
+    if isinstance(locale, str):
+        if locale not in Locale.values():
+            raise LocaleError(locale)
 
-    @classmethod
-    def values(cls):
-        return [i.value for i in cls.__members__.values()]
+        locale = Locale(locale)
+
+    if locale not in Locale:
+        raise LocaleError(locale)
+
+    return locale
