+++ mimesis/providers/address.py
@@ -40,6 +40,7 @@ class Address(BaseDataProvider):
         """Class for metadata."""
 
         name = "address"
+        data_provider = True
 
     @staticmethod
     def _dd_to_dms(num: float, _type: str) -> str:
