+++ mimesis/providers/finance.py
@@ -34,6 +34,7 @@ class Finance(BaseDataProvider):
         """Class for metadata."""
 
         name = "finance"
+        data_provider = True
 
     def company(self) -> str:
         """Get a random company name.
